+++ Changelog.md
@@ -4,6 +4,8 @@ All significant changes are listed in this file.
 
 ## [1.1.27] - 2023-4-28
 - Add automated tests for movie scrapers for improved future reliability
+- Fix several issues in JavDatabase actress scraper
+- Add filter to correct bad data in Actress database
 
 ## [1.1.26] - 2023-4-27
 - Fix multiple issues with actress scrapers